+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Not released
 
+- Search by coordinates supported in useGeocoderWidgetController hook [#731](https://github.com/CartoDB/carto-react/pull/731)
+
 ## 2.1
 
 ### 2.1.5 (2023-07-05)